Linx: - convert transcripts and fusion to file types - upload fusions and breakends from SV data loader

+++ hmf-common/src/main/java/com/hartwig/hmftools/common/variant/structural/annotation/FusionChainInfo.java
@@ -7,7 +7,7 @@ public abstract class FusionChainInfo
 {
     public abstract int chainId();
     public abstract int chainLinks();
-    public abstract long chainLength();
+    public abstract int chainLength();
     public abstract boolean validTraversal();
     public abstract boolean traversalAssembled();
 }
